feat (map) : add map on attraction board
명소 게시판에 위치 정보 추가

+++ src/main/java/com/cloud/web/service/AttractionService.java
@@ -79,8 +79,8 @@ public class AttractionService {
                 String uc_seq = String.valueOf(obj.getInt("UC_SEQ"));
                 String main_title = obj.getString("MAIN_TITLE");
                 String gugun_nm = obj.getString("GUGUN_NM");
-                String lat = String.valueOf(obj.getLong("LAT"));
-                String lng = String.valueOf(obj.getLong("LNG"));
+                String lat = String.valueOf(obj.getDouble("LAT"));
+                String lng = String.valueOf(obj.getDouble("LNG"));
                 String place = obj.getString("PLACE");
                 String title = obj.getString("TITLE");
                 String subtitle = obj.getString("SUBTITLE");
@@ -183,8 +183,8 @@ public class AttractionService {
                 String uc_seq = String.valueOf(item.getInt("UC_SEQ"));
                 String main_title = item.getString("MAIN_TITLE");
                 String gugun_nm = item.getString("GUGUN_NM");
-                String lat = String.valueOf(item.getLong("LAT"));
-                String lng = String.valueOf(item.getLong("LNG"));
+                String lat = String.valueOf(item.getDouble("LAT")); // todo : getDouble로 변경
+                String lng = String.valueOf(item.getDouble("LNG"));
                 String place = item.getString("PLACE");
                 String title = item.getString("TITLE");
                 String subtitle = item.getString("SUBTITLE");
